Fixed ids - changed to identity, hall bug fixed, show CRUD.

+++ Reservation_App/src/main/java/com/reservationapp/controller/ReservationController.java
@@ -102,6 +102,10 @@ public class ReservationController {
 	
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<Reservation> delete(@PathVariable Long id) {
+		Reservation res = reservationService.findOne(id);
+		if(res.getSeats().getHall().getInstitution().getAdmin() != loggedUser()){
+			return new ResponseEntity<>(res, HttpStatus.UNAUTHORIZED);
+		}
 		Reservation deleted = reservationService.delete(id);
 		return new ResponseEntity<>(deleted, HttpStatus.OK);
 	}
